drbd: Consider outdating a disk when more recent data is behind a diskless
Consider nodes A, B, and C, where B is intentionally diskless:

   B     | with B diskless and primary. It writes some data.
  / \    | B demotes to secondary.
 A---C   | A and C have backing disks.

   B     | C becomes isolated. B promotes, writes and demotes
  /      | again.
 A   C   | C is now an older data generation.

   B     | B and C establish a connection.
  / \    |
 A   C   |

This change is about the last step. While with the previous commit B
already marks C as Outdated, this commit introduces that C becomes
outdated while establishing the connection to B.

Consider also this scenario:

   B     | with B diskless and primary. It writes some data.
  / \    | B demotes to secondary.
 A---C   | A and C have backing disks.

   B     | C becomes isolated. B promotes, writes and demotes
  /      | again.
 A   C   | C is now an older data generation.

   B     | B down and up.
         |
 A   C   |

   B     | B and C establish a connection.
    \    |
 A   C   |

   B     | B and A establish a connection.
  / \    |
 A   C   |

This change introduces that B sets its exposed_data_uuid to the
current_uuid of the first UpToDate node it connects to.
When the connection with A gets established a moment later, it sets
C's disk to Outdated in the same state change.
Finally, so that C also notices that it should outdate itself, this
change also introduces that a node sends updates when it changes ist
exposed_data_uuid. When such an update is received, a node serving as
the data source for a diskless then outdates itself if necessary.