drbd: Make consistent neighbours of a diskless up-to-date when UUIDs match
The previous commit introduced that peers of a diskless are, in the
best case, considered as D_CONSISTENT as long as they are not on the
same UUID. When such a peer becomes D_UP_TO_DATE, since it finishes a
resync it sends its state first, then its new UUIDs.

When the diskless sees that this neighbor arrived at the same UUID,
elevate it to D_UP_TO_DATE if it is consistent.