drbd: only remove resync request on SyncSource from list once
The peer request is already removed from resync_ack_ee when it is moved
to the "matching" list in find_resync_requests(). So we do not need to
and should not try to remove it again in drbd_resync_read_req_mod().

This was harmless in the normal case where the peer request is freed in
the drbd_resync_read_req_mod() call from got_RSWriteAck(). It caused
problems in the case where the peer request is freed in the
drbd_resync_read_req_mod() call from drbd_rs_reply(). In this case,
got_RSWriteAck() has already run and has moved the peer request to a
list on the stack which no longer exists, so w.list must not be used.

The problems were seen in crashes like this:

list_del corruption. prev->next should be ffff90151d051e40, but was ffffffff8b049d8a
...
Call Trace:
 drbd_resync_read_req_mod+0x81/0xe0 [drbd]
 drbd_rs_reply+0x2a5/0x2f0 [drbd]
 w_e_end_rsdata_req+0x82/0x270 [drbd]

Fixes: 38ae3e9fb7f2 ("drbd: simplify peer request lists")