drbd: use del_timer for connect_timer, not timer_shutdown
The newly invented "shutdown" timer state indicates that the timer
should no longer be able to be re-armed because it will be freed soon.

Upon first glance, this usage of del_timer seemed like a candidate for
conversion to timer_shutdown_sync, but it turns out I missed another
usage in drbd_receiver.c, in conn_disconnect().

The relatively subtle but disastrous effect of this is that, once a
connection is disconnected, it can no longer be connected because the
connect_timer cannot be re-armed.

So, we need to use both variants. In drbd_unregister_connection(), the
timer actually goes away "for good", so use timer_shutdown_sync there.
In conn_disconnect(), keep using del_timer.

Fixes: 0aeea3432534 ("drbd: Convert del_timer*() to timer_shutdown*()")