build: enable gcov
To be able to collect code coverage data when running the DRBD test
suite.

Coverage collection is disabled by default, and can be enabled by
defining GCOV_PROFILE=y when building the module.

Route this through the rpm/deb packaging magic.
For rpm, --with-gcov can be passed to enable the behavior.
For debian, DEB_BUILD_OPTIONS must be set to "gcov".