drbd: Fix condition for clearing bitmap
When two nodes connect (or one side attaches a disk) and no resync
starts, we should remove the bitmap (and bitmap-uuid) between the two.

The bug was that this also cleared the bitmap if the peer is a
permanently diskless node. That was wrong.

It looks like an attempt to keep the bitmap-UUIDs to intentional
diskless peers all zero. It was wrong since we should keep that
bitmap-UUID as long as the local node has a bitmap towards this peer.

Fixes: 501071ca86ed "drbd: move bitmap clearing from strategy_to_repl_state() to w_after_state_change()"