drbd: restore strict protocol compatibility with last released drbd version
The recent commits 195d1273abb7 and 772e5b21d899 introduce that a
diskless node considers peers that are on a different UUID as
Consistent instead of UpToDate (when such a peer considers itself as
UpToDate).

To avoid issues while upgrading a cluster, enable that new behavior
only when the peer is on the same protocol level.