drbd: check connection state of an affecte_connection
When a node gets a two-phase-commit indirectly via another connection,
and it sees the affected connection as connected, the initiator of the
two-phase-commit likely has another view of the world.
The expectation is that the initiator would have sent the
two-phase-commit request directly.
Checking that affected_connection with a P_PING will probably change
its state to something lower than C_CONNECTED.

This is important so that such a node executes the far_away_change()
code further down and outdates itself.

The sequence of events that highlighted this bug are:

1: A cluster of 3 nodes. A and B with a backing disk, C diskless.
A is primary

  A---B
   \ /
    C

2: A has a request timeout towards B and closes to connection towards B.
(Note that B has not yet realized that the connection to A is down)
A starts a new-data-generation now, since it continues to write after
it closed the connection.

  A  -B
   \ /
    C

3: A demotes to secondary. C and B are part of that cluster-wide state
transition. B now knows that A is no longer a primary.

  A  -B
   \ /
    C

4: Finally, B realizes that the TCP connections to A no longer exist.

  A   B
   \ /
    C

5: Now, B is free to promote.

At the end of step 2, A triggers an empty two-phase-commit to test if
B is still reachable. With this fix, B will realize what happens and
outdate itself when committing that two-phase-commit.