drbd: propagate new current UUID to established peers
This is important so that our peers update peer_device->current_uuid
towards us. This has an effect when they update their UUIDs at the end
of a resync from us. For instance, at the end of a resync after a
resize.

Without this change, a resync after a resize can cause the current UUID
on the SyncTarget to jump back to an older UUID. One consequence of this
is that a later resync will be a full resync rather than a bitmap-based
one.