drbd: reset diskless peer slot to day0 UUID when forgetting
Fixes a bug with the same symptoms as and a similar cause to:
d8f5cab5b561 drbd: do not copy zero UUID from diskless peer when freeing peer slot

In this case the problem was that forgetting a diskless peer left a peer
slot in the metadata that looked like a day0 slot but contained a zero
UUID.

This change also has the effect on the metadata when we remove the
bitmap for a peer. That is, when a peer had a disk and becomes
intentionally diskless. Previously, we set the bitmap UUID for this
peer's slot to the day0 UUID. This is a little unexpected. Now we set
the bitmap UUID to zero. It will, of course, be set to the day0 UUID if
we later forget the peer.