drbd: mask out UUID_PRIMARY bit when checking whether peer moved on
It is possible for a diskless peer to send us a current UUID packet with
UUID_PRIMARY set followed by one with the same UUID, but UUID_PRIMARY
not set. The result was that all nodes with disk ended up in state
Outdated.

Mask out the UUID_PRIMARY bit to avoid this.