drbd_transport_tcp: add kTLS support
A kernel socket can be configured to use TLS, by offloading the handshake to a
user space utility. After the userspace performs the handshake, it configures
the socket with TLS in the kernel.

This requires some changes to the way the control socket is handled: since the
kTLS implementation does not support read_sock until kernel 6.6+, we can't use
the same path as in the normal TCP case. Even with kernel 6.6+, the read_sock
implementation is not safe to call in SOFTIRQ context.

Instead, we add a work to read messages from the control socket. The work is
triggered in the data_ready callback. We also need to temporarily disable
the data_ready callback while the socket is used by the user space handshake
utility.