drbd: ensure that resync work runs when states flap
It is possible for sync states to flap as below. Notice that the resync
towards d-150 starts, but is then paused because d-151 appears to be the
desired peer. But the resync to d-151 never starts. Instead d-150 is
chosen as the desired peer in the next state change and that connection
becomes SyncTarget again.

[   30.299134] drbd res/0 drbd1 d-150: repl( WFBitMapT -> SyncTarget )
[   30.301258] drbd res/0 drbd1 d-150: Began resync as SyncTarget (will sync 184320 KB [46080 bits set]).
...
[   30.737588] drbd res/0 drbd1 d-151: repl( WFBitMapT -> PausedSyncT ) resync-susp( connection dependency -> no )
[   30.740073] drbd res/0 drbd1 d-150: repl( SyncTarget -> PausedSyncT )
[   30.741976] drbd res/0 drbd1 d-150: Resync suspended
[   30.743946] drbd res/0 drbd1 d-151: Began resync as PausedSyncT (will sync 184320 KB [46080 bits set]).
[   30.796187] drbd res/0 drbd1 d-151: resync-susp( no -> connection dependency )
[   30.798806] drbd res/0 drbd1 d-150: repl( PausedSyncT -> SyncTarget )
[   30.801174] drbd res/0 drbd1 d-150: Syncer continues.

This can cause resync work to stop being scheduled. Using the names from
above:
* We are running our resync from d-150.
* Our replication state to d-150 changes to PausedSyncT.
* w_resync_timer() runs and does nothing because there are still
  outstanding resync requests. In particular, it does not re-arm the
  resync_timer.
* drbd_rs_all_in_flight_came_back() runs and schedules resync_work.
* w_resync_timer() runs and performs a state change to set resync_active
  to false. However, the state logic chooses d-150 as the desired peer
  again, so resync_active remains true.
* As a result, w_after_state_change() does not schedule resync_work.

Fix this by scheduling resync_work based on the change of replication
state rather than the change to resync_active.

Note that this always triggers when the previous code would have
triggered. A change of resync_active from false to true, means that the
peer was not the desired resync peer and now is, so it must also change
from a state other than SyncTarget, to SyncTarget.

Fixes: 61f1c6dd1328 ("drbd: serialize syncs from multiple sources")