drbd: clean up drbd_free_peer_req() and drbd_free_peer_req_no_list()
An API where you have two different functions to fee an object,
depending on whether that object is on some list, must be more
intuitive. It was like a pitfall that waits to get you.

It is easier to mark the peer_req object with a flag, indicating
whether it is on such a list.