drbd: Stop turning off local IRQs too long
This code snippet:

if (need_resched()) {
	spin_unlock_irq(&bitmap->bm_lock);
	cond_resched();
	spin_lock_irq(&bitmap->bm_lock);
}

May give the current CPU to other tasks that became runnable. However,
if there is no other runnable task, it blocks interrupts on the local
CPU potentially for a long time.

When the bitmap is large enough, and the network interrupt happens to
be routed to the CPU that executes one of these bitmap operations,
this might delay the processing of the network interrupt for so long
that DRBD reports a ping timeout.

We spotted this because of ping-timeouts, but it also affected other
parts of the system, even the timer interrupt incrementing the jiffies
counter.

Fix this by releasing the lock and enabling IRQs after each page. That
should result in an acceptable IRQ latency.