compat: test for lookup_user_key
When using TLS and configuring user provided key material, we make use of
lookup_user_key to ensure the user (the one calling drbdsetup) can actually
access the provided key.

Older kernels do not expose this, so we have to disable all code paths using
this function. This basically disables this particular feature of TLS.

This will most likely be fine, as there is still the fallback to configure
key material directly in the user space handshake utility.