drbd: build: fix spurious re-build attempt of compat.patch
Patching changed the timestamp of the patched files,
which are pre-requisites for generating the patch.

The second build after extracting the tarball
would try to regenerate the patch file.

Exclude generated *.mod.c from "to-be-patched" files for spatch.
Reset timestamp of patched files to that of the patch file, instead of current time.