compat: block: introduce holder ops
See also upstream Linux kernel commit
0718afd47f70 ("block: introduce holder ops")

Original message:

Add a new blk_holder_ops structure, which is passed to blkdev_get_by_* and
installed in the block_device for exclusive claims.  It will be used to
allow the block layer to call back into the user of the block device for
thing like notification of a removed device or a device resize.