lb-tcp: fix crash with unexpected connects and path removal
When a path gets removed, and concurrently, a connect attempt comes
in, it might happen this way:

remove_path()
  spin_lock_bh(&dtl_transport->paths_lock);
  list_del_init(&drbd_path->list);
  spin_unlock_bh(&dtl_transport->paths_lock);
  drbd_put_listener()
    path->listener = NULL
    spin_lock_bh(&listener->waiters_lock);
    list_del(&path->listener_link);
    spin_unlock_bh(&listener->waiters_lock);
                                                             *connect comes in
							     kref_get(listener)
							     schedule_work()...

    kref_put(&listener->kref, drbd_listener_destroy);
...

Then, del_accept_work_fn() could not find the path, and the list was
empty. So also, the return value of list_first_entry() is bogus.

Fix this mess by not trying to use a transport for logging purposes.