drbd_transport_tcp: fix read_sock cb read limits
When using skb_prepare_seq_read and skb_seq_read, we might get back a block
of not fully ready data. So we have to manually ensure we do not read past
the limit of ready data.

This does not seem to be a problem on the typical TCP stack, only if there is
some form of layered protocol (such as TLS) on top did this become an issue.
It is still good to be prepared, so ensure we apply useful bounds.