lb-tcp: Fix potential crash when unloading the transport module
When one downs a connection and soon after unloads the transport
module and, a connection gets accepted after the DTL_CONNECTING was
cleared in dtl_free().

Due to reference counting, the listener and path objects still
exist. The dtl_accept_work_fn() drops these references when it exits.

Then, it is a bad idea to add sockets to paths with socket callbacks
set to code that might become unavailable after dropping the last
reference on the listener object.