lb-tcp: Do not restore the socket callbacks when freeing sockets
Unregistering the callbacks races with an invocation of the callbacks
and restoring them to their original values. I.e., we get a call into
dtl_control_state_change(), and at the time that reaches the call to
original_sk_state_change, that pointer was set to NULL by
dtl_restore_socket_callbacks().