transport: add return code for net_conf_change
The transport might need to do additional validation on the net_conf
parameters. To support that, we add a return code: if negative, the
new net_conf is not applied.

This is done in preparation for TLS support, as there we need
additional (fallible) calls when updating the net_conf.

To ensure that the validation is always run, we also call net_conf_change
when a new connection is created: the trnasports can check this by observing
a NULL net_conf on the transport.