drbd: improve handling of "unexpected" write requests
Rename "drbd_fail_request_early" into "drbd_reject_write_early",
because that is what we attempt to do here:
fail write requests, if we are not currently in primary role.

There are various ways to end up with write requests,
even if we never saw a drbd_open() in write mode.

It is legal to call fsync() on a file descriptor that was opened O_RDONLY.
Some tools do that. If that fd is a block device,
that can cause an "empty flush" write request.

Which we can fail silently, there is no point in spamming the log
with scary messages for an fsync on a read-only fd.

It is legal to "mount -o ro", then later "mount -o remount,rw".
The read-only mount will call drbd_open() in read-only mode.
The remount will not call drbd_open at all.

The remount,rw will fail if at that point in time we are Secondary,
but it may succeed if we happen to be Primary.

But it was still counted as read-only opener,
so a later demote to Secondary would succeed,
and we would end up with a read-write mounted Secondary,

Any write requests from that mount would now suddenly be rejected.
Which is ... unexpected.

If we see a write-request while being Primary,
on a device that was not yet marked as "writable",
we now implicitly promote that device to "writable".
Its openers are now counted as writers, which prevents the demotion.

If we still see write requests to a Secondary, or to a device with
an open count of zero, we will still be noisy.
If we have opener information for the target device, we append it
to the "Rejected WRITE request ..." message;
unless we are seriously short on memory.