lb-tcp: only use first established control flow
Distributing the control packets over the available paths causes
issues on the receiving side. One of these issues is a bug in the
generic kernel that got fixed with the Linux kernel commit:

commit c457985aaa92e1fda2ce837cabf90bf687b92dcb
Author: Cong Wang <cong.wang@bytedance.com>
Date:   Wed Aug 17 12:54:43 2022 -0700

    tcp: fix tcp_cleanup_rbuf() for tcp_read_skb()

That commit comes with linux-6.0 and later.

For now, use only a single path for all control packets. The required
bandwidth on the control connection should be 100 times lower than
what we need on the data connection. It is not an issue using only one
for now.