drbd: soft-dependency on handshake
Traditionally DRBD loads the transport modules when it needs
them. I.e. TCP or RDMA (and soon also LB-TCP).
The recent commits introduced a hard module dependency from
drbd_transport_tcp to handshake. (The handshake module handles
handshakes for TLS.)

The handshake module takes the netlink lock in its module_init()
function. DRBD loads the required transport module in a netlink
callback. Netlink holds this lock when it invokes such callbacks. That
is a deadlock.

Working around this deadlock with a soft-dependency from DRBD to the
handshake module.