drbd_transport_tcp: fix tls key config
We used to store direct references to the keys provided by the user,
having to balance key_get() and key_put to keep the keys alive.

This introduced some counting bugs for very little benefit: we can simply
use the key_serial_t provided by the net_conf directly. Permissions are
then checked in the handshake module, so that tlshd only gets access to
keys it can see, or keyrings that can be linked to.