compat: fix double free in handshake module
This bug show itself as a null dereference in clean up code:

[ 4553.757168] BUG: kernel NULL pointer dereference, address: 0000000000000030
[ 4553.763289] #PF: supervisor read access in kernel mode
[ 4553.768699] #PF: error_code(0x0000) - not-present page
[ 4553.774039] PGD 0 P4D 0
[ 4553.776908] Oops: 0000 [#1] PREEMPT SMP NOPTI
[ 4553.779420] CPU: 0 PID: 1851 Comm: drbd_r_res1 Tainted: G           OE    --------  ---  5.14.0-284.30.1.el9_2.x86_64 #1
[ 4553.782988] Hardware name: QEMU Standard PC (Q35 + ICH9, 2009), BIOS 1.16.2-1.fc38 04/01/2014
[ 4553.785417] RIP: 0010:handshake_req_cancel+0xc/0xe0 [handshake]
[ 4553.787693] Code: db 74 0d 48 89 ef 48 89 d8 5b 5d ff e0 0f 1f 00 5b 5d c3 cc cc cc cc 0f 1f 80 00 00 00 00 0f 1f 44 00 00 41 54 55 48 89 fd 53 <4c> 8b 67 30 e8 ab fe ff ff 48 85 c0 0f 84 92 00 00 00 4c 89 e7 48
[ 4553.791948] RSP: 0018:ffffb447009a3cd8 EFLAGS: 00010286
[ 4553.793310] RAX: ffff955d474b7500 RBX: ffffb447009a3d50 RCX: 0000000005200000
[ 4553.794816] RDX: 0000000005000000 RSI: 0000000000000046 RDI: 0000000000000000
[ 4553.796351] RBP: 0000000000000000 R08: ffffffff95a74a88 R09: ffffffff95a74aa0
[ 4553.797918] R10: 0000000000000cf2 R11: 0000000000000002 R12: ffffb447009a3d90
[ 4553.799336] R13: ffff955d474b7500 R14: ffffb447009a3d48 R15: ffffb447009a3d50
[ 4553.800824] FS:  0000000000000000(0000) GS:ffff955d7ec00000(0000) knlGS:0000000000000000
[ 4553.802225] CS:  0010 DS: 0000 ES: 0000 CR0: 0000000080050033
[ 4553.803323] CR2: 0000000000000030 CR3: 0000000007a26002 CR4: 0000000000370ef0
[ 4553.804629] Call Trace:
[ 4553.805440]  <TASK>
[ 4553.806218]  dtt_socket_free+0x1a/0x60 [drbd_transport_tcp]
[ 4553.807323]  dtt_connect+0x7cd/0xd53 [drbd_transport_tcp]
[ 4553.808402]  conn_connect+0x8e/0x5e0 [drbd]
[ 4553.809365]  drbd_receiver+0x14/0x60 [drbd]
[ 4553.810321]  drbd_thread_setup+0x74/0x1e0 [drbd]
[ 4553.811319]  ? drbd_reclaim_path+0x90/0x90 [drbd]
[ 4553.812356]  kthread+0xd6/0x100
[ 4553.813226]  ? kthread_complete_and_exit+0x20/0x20
[ 4553.814200]  ret_from_fork+0x1f/0x30
[ 4553.815070]  </TASK>

The issue being: we are in the netlink callback for the handshake user space
component. We install the sock as file descriptor, then try to create the
netlink message. If there is a problem creating the netlink message, we try
to clean up the installed socket file.

Using "fput(sock->file)" decremented the reference count, but left the FD in
the files table. Presumably shortly afterwards, the user space process will
exist upon observing the netlink error. This will also again close all FDs,
including the sock->file we already decremented. So after incrementing the
count in handshake_dup() once, we effectively decrement it twice, causing
the reference count to become unbalanaced.

This later exploded in our own clean up of the socket, because it was already
cleaned up.

The fix is to only install the file at the FD as a last step. Instead of fput()
we simply put the unused fd in the error case.