drbd_transport_tcp: normalize TLS error returns
The TLS handshake utility may promise to return negative values for errors,
but does not actually guarantee it anywhere: The user space utility could
return any int value.

This leads to issues, since we often use "err < 0" during the connect phase
to detect errors. So we ensure that the TLS errors are always negative.