lb-tcp: simplify dtl_send_zc_bio()
This is a non-functional change. The deep indentation in
dtl_send_zc_bio() made it hard to follow. Factor out two functions to
make it clearer.

The motivation for this change is to make the spatch step in the build
fast again. Prior to the introduction of lb-tcp, spatch typically took
less than 1 minute. Since then, spatch has taken about 2 minutes longer
for certain kernels. In particular, this cocci transformation from
struct_bvec_iter__no_present.cocci was slow:

@@ identifier bvec, f; @@
- struct bio_vec bvec;
+ struct bio_vec *bvec;
<...
- bvec.f
+ bvec->f
...>