Revert "drbd: improve merging of write acks into larger packets"
This reverts commit d1e2c30c1481d94f512ff966379a0525021c28ae.

This commit causes writes to sometimes take 10s. That is, until the next
ping. This is fixable. For instance, with one of the following fixes:
* Flush the control stream when sending barrier acks
* Move the queuing of send_acks_work in drbd_endio_write_sec_final()
  below the update to active_ee_cnt.

However, even with these fixes the change causes a performance
regression in a different situation. Specifically, when an application
is sensitive to write completion latency. Consider a setup when a
diskless Primary and a Secondary with a disk that completes requests
after either 5ms or 15ms. This could be 2 volumes in one resource. The
application is writing 2 streams, and pauses 12ms between each write.

With d1e2c30c1481, the following occurs (the prefix is the timestamp in
ms):
0 Secondary submits write W1
0 Secondary submits write W2
5 W1 completes, no write-ack sent due to d1e2c30c1481
15 W2 completes, write-acks sent for W1 and W2
27 Application submits next writes

Without d1e2c30c1481, the following occurs:
0 Secondary submits write W1
0 Secondary submits write W2
5 W1 completes, write-ack sent
15 W2 completes, write-ack sent
17 Application submits write W3 following after W1
17 Secondary submits write W3
22 W3 completes, write-ack sent

We see that, without d1e2c30c1481, an additional write occurs in the
same timeframe. A performance regression of 20% was measured in one
experiment as a result.