lb-tcp: set bit DTL_LOAD_BALANCE in dtl_net_conf_change()
dtl_init() is called before the net conf is set, so it is not an
appropriate place to check load_balance_paths.

DRBD prevents changes to load_balance_paths because it is marked as an
invariant flag. So this will not cause DTL_LOAD_BALANCE to be set after
the handshake has started.