dkms: automatically guess if we should build against OFED
Before, you would have to explicitly "export BUILD_OFED=1" first.
Now, we detect an OFED build in dkms.conf, then also guess which
of the potentially many source links below /usr/src/ofa_kernel/*/*
to use (with a fallback to ".../default").

If that should happen to guess wrong sometimes
for example you currently run with ofed enabled,
but you mean to build for ofed disabled,
export BUILD_OFED=0 (or anything that is not the empty string or 1).