drbd_transport: fix regression from consolidating function pointers
The commit b426d13cab89 removed the generic_listener_destroy()
function. Now, the listener_destroy() functions of the transports need
to deal with objects on which the init_listener() returned an error,
and did not initialize the private members.

In case of the drbd_transpot_tcp the s_listen member.