WinDRBD: Fix for busy write resync stall
When an application writes the same block over and
over again, sync cannot proceed and therefore never
finishes. WinDRBD has the ability to suspend application
I/O from outside DRBD. This patch adds logic to detect
sync stalls because of application I/O and calls to
the exported API functions of WinDRBD to suspend
I/O (when there is no sync progress within 5 seconds)
and resume I/O (when there is sync progress, the
sync finishes normally or the sync was aborted,
for example due to connection loss).