drbd: preserve quorum over a reboot
In a cluster with an even number of diskful nodes, a partition having
exactly half of the nodes with backing disks can keep quorum when it had
contact to the majority of the diskless nodes at the moment when
diskful node count dropped to 50%.

Here is an example with two diskful nodes and with B being the diskless
tiebreaker:

   B        B
  / \  =>  /    =>   => [ ] =>
 A---C	  A  	   A           A

This change introduces that when A reboots (or crashes), it restores
its quorum status when coming up again. The reasoning is that the
other nodes can not gained quorum.

Handling clusters of five (4 disk nodes with one tiebreaker) is more
involving. It would require that the nodes compare their prev_members
masks.

This commit enables the restoration of quorum in 3-node clusters only.