drbd: fix connecting nodes with differently sized backing disks
Consider a cluster of 3 nodes with all possible connections
configured. The only obstacle is that one node (A) has a slightly
smaller backing disk, and it is primary before establishing the
connections.

       B
      / \
     A---C

When the nodes connect in this order:

 1. A and B connect, agreeing on the smaller size.
 2. B and C connect and agree on the larger size.
    Then B sends that larger size to A, and A reacts by dropping
    the connection as it can not fulfill that size requirement.

I believe this behavior goes back to the two node times of DRBD. I
solve this by changing the meaning of the d_size field of the p_sizes
packet. Instead of sending the user-data capacity of the local
backend, send the minimum size of the user-data capacity of the
visible cluster partition (including my backing disk, if it exists).