drbd: fix size propagation
In a linear chain of nodes

 A - B - C

When A joins last and has a smaller backing disk than B and C, the
smaller size also needs to propagate to C. DRBD has code for that
propagation in receive_sizes(), expressed by the boolean
should_send_sizes.

That boolean gets set when the size changes. The bug was that
drbd_determine_dev_size() returned unexpectedly DS_GREW_FROM_ZERO.

I am fixing that by comparing the current size to the previous size.