drbd: modernize failure mode when disk sizes do not match
Consider nodes a linear chain of nodes:   A - B - C .

B's and C's backing disks have the same size, and A's is smaller.

B and C connect first, and C becomes primary. The A connects. This
fails as it should, but so far, B closed the connection during the
handshake. That led to a two-phase commit timeout.

I am improving that by doing a graceful failure of the
two-phase-commit, and both nodes transition to StandAlone.