drbd: Fix committing handshake result
The following log shows node st05 that should become sync-source
towards node st03. It does a handshake with the expected result.

But while it tries to commit this change into its state, the handshake
code executes one more time and overwrites the result that should be
committed.

15:51:46.673686+01:00 vm-132-disk-1: Preparing cluster-wide state change 3696953758 (0->1 499/146)
15:51:46.709662+01:00 vm-132-disk-1/0 drbd1034 st03: drbd_sync_handshake:
15:51:46.709681+01:00 vm-132-disk-1/0 drbd1034 st03: self 3DEAC516805C3E1E:F6D6F2955A6080C8:0000000000000000:0000000000000000 bits:0 flags:20
15:51:46.709683+01:00 vm-132-disk-1/0 drbd1034 st03: peer 0000000000000004:0000000000000000:F6D6F2955A6080C8:0000000000000000 bits:525190 flags:24
15:51:46.709684+01:00 vm-132-disk-1/0 drbd1034 st03: uuid_compare()=source-set-bitmap by rule=just-created-peer
15:51:46.709685+01:00 vm-132-disk-1/0 drbd1034 st03: Setting and writing one bitmap slot, after drbd_sync_handshake
15:51:46.713633+01:00 vm-132-disk-1/0 drbd1034 st03: updated bitmap UUID 3DEAC516805C3E1E:0000000000000004:0000000000000000:0000000000000000
15:51:46.738518+01:00 vm-132-disk-1: State change 3696953758: primary_nodes=0, weak_nodes=0
15:51:46.738536+01:00 vm-132-disk-1: Committing cluster-wide state change 3696953758 (64ms)
15:51:46.738537+01:00 vm-132-disk-1/0 drbd1034 st03: drbd_sync_handshake:
15:51:46.738538+01:00 vm-132-disk-1/0 drbd1034 st03: self 3DEAC516805C3E1E:0000000000000004:0000000000000000:0000000000000000 bits:0 flags:20
15:51:46.738540+01:00 vm-132-disk-1/0 drbd1034 st03: peer F6D6F2955A6080C8:0000000000000000:F6D6F2955A6080C8:0000000000000000 bits:525190 flags:24
15:51:46.738542+01:00 vm-132-disk-1/0 drbd1034 st03: My bitmap UUID changed during handshake. 0xF6D6F2955A6080C8 to 0x4
15:51:46.738550+01:00 vm-132-disk-1/0 drbd1034 st03: uuid_compare()=retry-connect by rule=initial-handshake-changed
15:51:46.753643+01:00 vm-132-disk-1 st03: conn( Connecting -> Connected ) peer( Unknown -> Secondary ) [connected]
15:51:46.753650+01:00 vm-132-disk-1/0 drbd1034 st03: pdsk( DUnknown -> Inconsistent ) repl( Off -> Established ) [connected]

This is a race condition. In the above log, the peer sends another
UUID and STATE packet exceptionally quickly.

I am fixing this issue by making sure we execute the handshake code
only once when establishing the connection.

When the passive side (the node with the higher node-id) rejects the
two-phase commit, it will receive the uuid and state packet
"unexpectedly". The active side retries the state change and sends it
again after the P_TWOPC_PREPARE packet.
An earlier version of this patch broke this scenario because it only
prevented double processing of UUID and STATE. This final version also
disregards the "unexpected" UUID and STATE packages arriving too early
because of an aborted two-phase commit.