docker: find matching RPM based on metadata
For RHEL-like systems, our RPMS contain the expected symbol versions
in the package metadata. We can use this information to find the RPM
matching the currently running kernel: We invoke the scripts used
to generate these requirements on the currently running kernel and
and match them to the available RPMs.

This was triggered by a breakage in compatibility of a RHEL9 kernel,
where the matching OL9 kernel stayed compatible with the older version.
In this case the lbdisttool.py failed to generate the right module
to load on OL9. With this change, the correct modules for RHEL9 and OL9
are loaded.