drbd: fix aborting two-phase-commits returning to UpToDate
In a cluster of three, when the primary gets isolated, the two
remaining secondaries downgrade their disks to Consistent and try to
return to UpToDate as soon as possible.

They use a two-phase-commit with an unspecified target node (a
target_node_id value of -1). They might prepare that two-phase-commit
before the receiver calls conn_disconnect() on the connection to the
former primary.

So far, DRBD has failed to abort such a two-phase-commit promptly
since drbd_twopc_between_peer_and_me() did not consider that the
target_node_id value might be -1.