drbd: Improve resync decision on the second connection
Consider the scenario:

    B    | 1. Resync starts from C to B while connecting
   / \   | 2. B connects with A, marking it a PausedSyncS
  A   C  |

When A and C are on a newer current-UUID than B, DRBD decided for a
partial resync at event 2.

When A, B, and C are on the same UUID, e.g., because of a common power
failure, DRBD wrongly decided for a full resync.

So far, the functions drbd_disk_states_source_strategy() and
drbd_disk_states_target_strategy() only considered the cases when the
resync target node is at a different current UUID as the sync source
nodes.

I am fixing that suboptimal decision in these two functions for the
described scenario.

This scenario can temporarily appear when a cluster starts up and was
the source for unexpected full resyncs.