drbd: fix potential module refcount "leak"
If we have some openers on a primary device,
then "secondary --force" and "down" a resource,
the last "release" will trigger a state change
on the "zombie" resource:
    force-io-failures( yes -> no ) [release]

A state change allocates and queues an after-state-change work,
to be consumed by the resource's worker thread -- which at that time
is already gone.
Now the reference counts associated with that work item are pinning the
"zombie" resource, which is pinning the module:
we can no longer unload the module.

To fix that, I chose to not allocate an after state change work item
if the resource is already "unregistered": a non-existing work item
can be not processed by a non-existing worker thread just fine,
and does not hold any refcounts either.