build: explicitly list to-be-processed files for coccinelle / spatch
Using make $(wildcard *.c) and similar may pick up generated or
other untracked files. The resulting patch may contain chunks
for these files, which won't be present in a clean checkout or
fresh tarball extract. That would cause the build to fail, because
the compat patch does not apply.

Explicitly enumerate which files we want to be processed by spatch
in a new make include file.

For our convenience, add some magic that reminds us to update the list
should we add or remove source files (in the checked locations).