No timeout waiting for becoming standalone in try_disconnect()
When we run drbdsetup disconnect followed by a del_path it
may happen that the receiver is still in the connect function
of the OS'es connect() routine. When this routine returns
data structures are referenced that already have been freed.

To reproduce it:

1) have a connected resource
2) run disconnect on the peer (so resource goes into Connecting)
3) change network config (for example the local port number)
4) run drbdadm adjust

Do this several (100+) times and (on WinDRBD) you run into
BSOD's, system hangs or similar effects.

This patch fixes the issue for the WinDRBD version of DRBD
9.0.32.